# docs/swagger.yaml
openapi: "3.0.2"
info:
  title: traP knoQ
  description: |
    This is a sample knoQ server.
  version: 1.2.0
servers:
  - url: http://ec2-54-238-146-214.ap-northeast-1.compute.amazonaws.com/api
    description: Development server
...
paths:
...
  /events/{eventID}:
    get:
      tags:
        - events
      summary: 一件取得
      operationId: getevent
      parameters:
        - name: eventID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          $ref: "#/components/responses/event"
    put:
      tags:
        - events
      summary: 部屋の使用宣言を更新
      description: adminsのみ
      operationId: updateevents
      parameters:
        - name: eventID
          in: path
          description: 変更する情報のID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        $ref: "#/components/requestBodies/PutEvent"
      responses:
        200:
          $ref: "#/components/responses/event"
        403:
          description: Forbidden
        400:
          description: Bad Request
    delete:
      tags:
        - events
      summary: 使用宣言を削除
      description: adminsのみ
      operationId: deleteRevents
      parameters:
        - name: eventID
          in: path
          description: 削除する情報のID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          $ref: "#/components/responses/Nocontent"
        403:
          description: Forbidden
        404:
          description: Not Found
...
components:
  schemas:
    UUIDs:
      type: array
      description: UUIDの配列
      items:
        type: string
        format: uuid
...
    event:
      type: object
      properties:
        name:
          type: string
          example: 第n回進捗回
        descirption:
          type: string
          example: 第n回の進捗会です。
        sharedRoom:
          type: boolean
          example: false
        timeStart:
          type: string
          example: "2004-04-01T12:00Z"
        timeEnd:
          type: string
          example: "2004-04-01T12:00Z"
        roomId:
          type: string
          format: uuid
          example: "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
        groupId:
          type: string
          format: uuid
          example: "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
        admins:
          $ref: "#/components/schemas/UUIDs"
    PutEvent:
      allOf:
        - $ref: "#/components/schemas/event"
      properties:
        tags:
          type: array
          items:
            $ref: "#/components/schemas/ReqTagNameAuth"
...
    ResponseEvent:
      allOf:
        - type: object
          properties:
            eventId:
              type: string
              format: uuid
              example: "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
        - $ref: "#/components/schemas/event"
        - type: object
          properties:
            place:
              type: string
            tags:
              type: array
              items:
                $ref: "#/components/schemas/ResponseTag"
            createdBy:
              type: string
              format: uuid
            createdAt:
              type: string
              example: "2019-11-02T15:24:42.5438806Z"
            updatedAt:
              type: string
              example: "2019-11-02T15:24:42.5438806Z"
            deletedAt:
              type: string
              format: omitempty
              nullable: true
...
    ReqTagName:
      type: object
      properties:
        name:
          type: string
          example: Vue
...
    ReqTagNameAuth:
      allOf:
        - $ref: "#/components/schemas/ReqTagName"
      properties:
        locked:
          type: boolean
          example: true
...
    ResponseTag:
      type: object
      properties:
        tagId:
          type: string
          format: uuid
          example: "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
        name:
          type: string
          example: vue
        locked:
          type: boolean
          example: true
...
  requestBodies:
...
    PutEvent:
      description: 予約の編集
      required: true
      content:
        application/json:
          schema:
            oneOf:
            - $ref: "#/components/schemas/PutEvent"
            - type: object
              properties:
                name:
                  type: string
                descirption:
                  type: string
                sharedRoom:
                  type: boolean
                timeStart:
                  type: string
                timeEnd:
                  type: string
                place:
                  type: string
                groupId:
                  type: string
                  format: uuid
                admins:
                  $ref: "#/components/schemas/UUIDs"
                tags:
                  type: array
                  items:
                    $ref: "#/components/schemas/ReqTagNameAuth"
...
  responses:
    Nocontent:
      description: Nocontent
...
    event:
      description: successful operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseEvent"
...